backend/requirements-compatible.txt
# Core API packages
fastapi>=0.95.0,<0.105.0
uvicorn>=0.20.0,<0.25.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.10.0,<3.0.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration
python-dotenv>=0.21.0

# Database
motor>=3.1.0

# Basic ML packages
numpy>=1.22.0,<1.25.0
pillow>=9.0.0
scikit-learn>=1.0.0
scipy>=1.8.0
tqdm>=4.64.0

# Comment out these packages for initial setup
# opencv-python-headless>=4.5.0
# opencv-contrib-python-headless>=4.5.0
# transformers>=4.26.0
# sentence-transformers>=2.2.0
# torch>=1.12.0
# torchvision>=0.13.0
# faiss-cpu>=1.7.0
# clip-by-openai==1.0
